pyproject.toml
[tool.poetry]
name = "PR-event"
version = "1.0.0"
description = "An open-source GitHub app for malware-in-code scanning"
authors = ["Matan Giladi <matan.giladi@apiiro.com>"]
license = "MIT"
repository = "https://github.com/apiiro/pr-event.git"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9.2"
cryptography = "44.0.0"
flask = "3.1.0"
gunicorn = "23.0.0"
pygithub = "2.5.0"
pyjwt = "2.10.1"
pygments = "2.19.0"
requests = "2.32.3"
semgrep = "1.102.0"
werkzeug = "3.1.3"
hvac = "2.3.0"
pytest = "8.3.4"

[tool.poetry.dev-dependencies]
pytest = "8.3.4"

[build-system]
requires = ["poetry-core=1.0.0"]
build-backend = "poetry.core.masonry.api"